Add files via upload
First instance of <about> and <index> pages (manifesto and landing page). These can be committed directly to the master branch as they are the first instance. Any future edits will be created in a new branch and then will be reviewed by Serena and Maria-Elena via a pull request before being being merged with the master branch.