potty-humour-index-contact
if you click on the 'potty humour' a href on the index page and the contact page, you will get a toilet-themed joke. It's an ugly pop up but it's all I can do (sorry!).

I still need to work on the contact form.

I created a javascript file but I haven't needed it yet (probably should've put my js in there but I DIDN'T!!!!)